+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Fix ["conflicting weak extern definition" error](https://github.com/rust-lang/rust/issues/85461) on windows.
+
 ## [0.1.10] - 2021-10-24
 
 - Fix a compatibility issue with `cc`. ([#98](https://github.com/taiki-e/cargo-llvm-cov/pull/98))